worked on updating engine

+++ src/ae/atlas.cpp
@@ -28,7 +28,7 @@ _Atlas::_Atlas(const _Texture *Texture, const glm::vec2 &Size, float Padding) :
 	TextureSizeInTexels(Size / (glm::vec2)Texture->Size),
 	Padding(Padding) {
 
-	Columns = Texture->Size.x / Size.x;
+	Columns = (uint32_t)(Texture->Size.x / Size.x);
 }
 
 // Destructor
@@ -36,7 +36,7 @@ _Atlas::~_Atlas() {
 }
 
 // Returns coords given a texture index
-glm::vec4 _Atlas::GetTextureCoords(int Index) const {
+glm::vec4 _Atlas::GetTextureCoords(uint32_t Index) const {
 	float X = Index % Columns;
 	float Y = Index / Columns;
